Crate nargo_toml

Modules

deserializers 🔒

Includes custom serde deserializers for Nargo.toml types that could be implemented using serde(untagged) but are not because using that leads to very poor error messages.

errors 🔒

flock 🔒

git 🔒

semver 🔒

Structs

NargoToml

Tracks the root_dir of a Nargo.toml and the contents inside the file.

PackageConfig

PackageMetadata

WorkspaceConfig

Enums

Config

Contains all the information about a package, as loaded from a Nargo.toml.

DependencyConfig

Enum representing the different types of ways to supply a source for the dependency

ManifestError

Errors covering situations where a package is either missing, malformed or does not pass semver validation checks.

PackageSelection

Functions

find_file_manifest

Searches for a Nargo.toml file in the current directory and all parent directories. For example, if the current directory is /workspace/package/src, then this function will search for a Nargo.toml file in

find_file_root

Returns the PathBuf of the directory containing the Nargo.toml by searching from current_path to the root of its Path, returning at the innermost directory found, i.e. the one corresponding to the package that contains the current_path.

find_package_manifest

Returns the PathBuf of the Nargo.toml file by searching from current_path and stopping at root_path.

find_package_root

Returns the PathBuf of the directory containing the Nargo.toml by searching from current_path to the root of its Path, returning the topmost directory found, i.e. the one corresponding to the entire workspace.

find_root

Returns the PathBuf of the directory containing the Nargo.toml by searching from current_path to the root of its Path. When workspace is true it returns the topmost directory, when false the innermost one.

get_package_manifest

Returns the PathBuf of the Nargo.toml file in the current_path directory.

path_root 🔒

Get the root of path, for example:

read_toml

Attempts to read the file at the provided toml_path as a Nargo.toml file, returning it if the read was successful.

resolve_package_from_toml 🔒

Resolves a Nargo.toml file into a Package struct as defined by our nargo core.

resolve_workspace_from_fixed_toml

Resolves a Nargo.toml _ into a Workspace struct as defined by our nargo core.

resolve_workspace_from_toml

Resolves a Nargo.toml file into a Workspace struct as defined by our nargo core.

toml_to_workspace 🔒