nargo_toml::errors
Enum ManifestError

pub enum ManifestError {

    MissingFile(PathBuf),
    ReadFailed(PathBuf),
    MissingParent,
    MissingPackageType(PathBuf),
    InvalidPackageType(PathBuf, String),
    MalformedFile(Error),
    UnexpectedWorkspace(PathBuf),
    MissingEntryFile {
        toml: PathBuf,
        entry: PathBuf,
    },
    MissingDefaultEntryFile {
        toml: PathBuf,
        entry: PathBuf,
        package_type: PackageType,
    },
    InvalidPackageName {
        toml: PathBuf,
        name: String,
    },
    InvalidDependencyName {
        toml: PathBuf,
        name: String,
    },
    InvalidDirectory {
        toml: PathBuf,
        directory: PathBuf,
    },
    GitError(String),
    MissingSelectedPackage(CrateName),
    MissingDefaultPackage(PathBuf),
    BinaryDependency(CrateName),
    MissingNameField {
        toml: PathBuf,
    },
    NoCommonAncestor {
        root: PathBuf,
        current: PathBuf,
    },
    SemverError(SemverError),
    CyclicDependency {
        cycle: String,
    },
    ParseExpressionWidth(String),

}

Errors covering situations where a package is either missing, malformed or does not pass semver validation checks.

Variants

MissingFile(PathBuf)

Package doesnโt have a manifest file

ReadFailed(PathBuf)

MissingParent

MissingPackageType(PathBuf)

InvalidPackageType(PathBuf, String)

MalformedFile(Error)

Package manifest is unreadable.

UnexpectedWorkspace(PathBuf)

MissingEntryFile

Fields

toml: PathBuf

entry: PathBuf

MissingDefaultEntryFile

Fields

toml: PathBuf

entry: PathBuf

package_type: PackageType

InvalidPackageName

Fields

toml: PathBuf

name: String

InvalidDependencyName

Fields

toml: PathBuf

name: String

InvalidDirectory

Fields

toml: PathBuf

directory: PathBuf

GitError(String)

Encountered error while downloading git repository.

MissingSelectedPackage(CrateName)

MissingDefaultPackage(PathBuf)

BinaryDependency(CrateName)

MissingNameField

Fields

toml: PathBuf

NoCommonAncestor

Fields

root: PathBuf

current: PathBuf

SemverError(SemverError)

CyclicDependency

Fields

cycle: String

ParseExpressionWidth(String)