nargo_toml
Function find_file_root

pub fn find_file_root(current_path: &Path) -> Result<PathBuf, ManifestError>

Returns the PathBuf of the directory containing the Nargo.toml by searching from current_path to the root of its Path, returning at the innermost directory found, i.e. the one corresponding to the package that contains the current_path.

Returns a ManifestError if no parent directories of current_path contain a manifest file.