nargo_toml

Function read_toml

pub fn read_toml(toml_path: &Path) -> Result<NargoToml, ManifestError>

Attempts to read the file at the provided toml_path as a Nargo.toml file, returning it if the read was successful.

Errors

ManifestError::ReadFailed if the file could not be read.

ManifestError::MissingParent if the Nargo.toml file did not have a parent directory.

[toml::de::Error] if the Nargo.toml file could not be converted to a valid Config.